Add type-specific get_children
Like nodes_of_class, get_children is elegant and flexible and slow.

    def get_children(self):
        for field in self._astroid_fields:
            attr = getattr(self, field)
            if attr is None:
                continue
            if isinstance(attr, (list, tuple)):
                for elt in attr:
                    yield elt
            else:
                yield attr

It iterates over a list, dynamically accesses attributes, does null
checks, and does type checking. This function gets called a lot, and
all that extra work is a real drag on performance.

In most cases there isn't any need to do any of these checks. Take an
Assign node for instance:

    def get_children(self):
        for elt in self.targets:
            yield elt

        yield self.value

It's known in advance that Assign nodes have a list of targets and a
value, so just yield those without checking anything.

This commit does not include a get_children for the Arguments class.
That class has a lot of "astroid fields" of various types, so it's
probably worth it to take slight performance hit to keep the code
cleaner. Or maybe not.

Some nodes don't have any "astroid fields". Those nodes have been
given the following get_children:

    def get_children(self):
        return (_ for _ in ())

This is somewhat ugly. Are there alternatives? using `yield from`
would be nice, but it's a syntax error in legacy Python.

    def get_children(self):
        yield from ()

It's possible to do a `yield` after `return`, but Pylint deems this
"unreachable code":

    def get_children(self):
        return
	yield