Add accept_visitor
There is a clever pair of mutually recursive functions for applying
transforms to nodes:

    def _visit(self, node):
        if hasattr(node, '_astroid_fields'):
            for field in node._astroid_fields:
                value = getattr(node, field)
                visited = self._visit_generic(value)
                setattr(node, field, visited)
        return self._transform(node)

    def _visit_generic(self, node):
        if isinstance(node, list):
            return [self._visit_generic(child) for child in node]
        elif isinstance(node, tuple):
            return tuple(self._visit_generic(child) for child in node)

        return self._visit(node)

This is an elegant way to deal with transforming any kind of node, but
as usual, this elegance is slow. _visit has to dynamically access an
attribute, then pass it to _visit_generic, which has to perform two
separate type checks, the _visit has to dynamically set the attribute.
As discussed in previous commits, performance can be improved by
exploiting the fact that all the attribute that need to be transformed
are known in advance.