Add support for more advanced inference for typing.NamedTuple
typing.Namedtuple are collections.namedtuple in disguise. We do quite a lot of
work around understanding namedtuples, but we weren't doing the same for typing.NamedTuple.
In order to get a consistent experience, we moved typing.NamedTuple inference support
in the brain file corresponding to collections's namedtuple, and there we can leverage the
namedtuple inference rewriting functions. Now a new capability that typing.NamedTuple has is
to understand that we can define methods on the class, so that they can be used on the instance
when instantiated.
Another small bug that was fixed was to put Name nodes in .bases when building a namedtuple,
instead of putting the class node itself.

Close PyCQA/pylint#1628